+++ RaffleRunner/GlobalUsings.cs
@@ -16,10 +16,11 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 // 
 // Created on     07/25/2022 @ 18:44
-// Last edited on 07/25/2022 @ 19:11
+// Last edited on 07/26/2022 @ 01:08
 #endregion
 
 global using System;
+global using Serilog;
 global using System.Net;
 global using System.Text;
 global using System.Linq;