+++ RaffleRunner/Program.cs
@@ -16,10 +16,13 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 // 
 // Created on     07/25/2022 @ 19:15
-// Last edited on 07/25/2022 @ 23:29
+// Last edited on 07/26/2022 @ 02:01
 #endregion
 
+using Serilog.Core;
+using Serilog.Exceptions;
 using RaffleRunner.Commands;
+using Serilog.Sinks.SystemConsole.Themes;
 
 namespace RaffleRunner;
 
@@ -28,15 +31,28 @@ namespace RaffleRunner;
 [Subcommand(typeof(CheckUpdatesCommand))]
 internal class Program
 {
+    public static readonly LoggingLevelSwitch LogLevelSwitch = new();
+    
     private static async Task<int> Main(string[] args)
     {
         Console.OutputEncoding = Encoding.UTF8;
+
+        Log.Logger = new LoggerConfiguration()
+                     .Enrich.WithExceptionDetails()
+                     .WriteTo.Console(
+                         theme: AnsiConsoleTheme.Code,
+                         outputTemplate: "[{Timestamp:HH:mm:ss}] [{SourceContext}] {Level:u3} {Message:lj}{NewLine}{Exception}"
+                     )
+                     .MinimumLevel.ControlledBy(LogLevelSwitch)
+                     .CreateLogger();
         
         AnsiConsole.Write(new Rule($"[#06b6d4]{GlobalShared.ProgramIdentifier}[/] by depthbomb").LeftAligned());
 
         int exitCode = await CommandLineApplication.ExecuteAsync<Program>(args);
         
         AnsiConsole.Write(new Rule());
+
+        Log.CloseAndFlush();
         
         return exitCode;
     }