+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.1.0.0 Beta
+## 0.2.0.0 β
+
+- Implemented new logging library
+- Fixed `check-updates` command failing
+
+## 0.1.0.0 β
 
 - Beta release